fix: Hold's Duration is int

+++ JipperResourcePack/PlayCount.cs
@@ -140,7 +140,7 @@ public class PlayCount {
                 case LevelEventType.Hold:
                     memoryStream.WriteInt(levelEvent.floor);
                     memoryStream.WriteByte(2);
-                    memoryStream.WriteFloat((float) levelEvent["duration"]);
+                    memoryStream.WriteFloat((int) levelEvent["duration"]);
                     break;
                 case LevelEventType.MultiPlanet:
                     memoryStream.WriteInt(levelEvent.floor);
